requirements.txt
fastapi
uvicorn
openai
chromadb
python-docx
requests
